requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Audio/Video Processing
openai-whisper==20231117
ffmpeg-python==0.2.0
moviepy==1.0.3

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# Quality Assessment
jiwer==3.0.3

# Translation
deep-translator==1.11.4
langdetect==1.0.9

# Utilities
httpx==0.25.2
rich==13.7.0

# System Monitoring
psutil==7.0.0

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
